General translation done.
So I finished translating episode 6 in general, but I missed 2 things that I didn't understand in Japanese so I might come back to this later. I hope you can manage fixing the timings because they're hella broke! I fixed the timing in the first 2-4 minutes or something but you'll notice that the subs are off sync as soon as you go past that. I think the honorifics in this episode are done as well, so you don't have to worry about that. I checked the entire episode and made sure there is nothing I missed. All that's left for you to do on this episode is to fix the timings and signs.